move file_paths function to root from featurizer
